Prepare code for WRS calls.
